refactor uci options
for some reason UciCommand::Stop does not abort immediately after the refactor. I have no idea what could cause this as the search still happens on a different thread than the input, and you can still do the input even while the search is happening, but the search will not abort immediately. I literally copy pasted the code from main.rs and modified it. So confused.